requirements.txt
django
 djangorestframework
 confluent_kafka
